fix: Modificacion script para unificar tareas
Con el fin de unificar tareas y gracias a la Programacion orientada a objetos se puedo juntar los scripts de populate_dim_user y el ETL
